Market Research Agent/requirements.txt
crewai>=0.11.0,<0.12.0
langchain>=0.1.0,<0.2.0
pandas>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
google-generativeai>=0.3.0,<0.4.0
newsapi-python>=0.2.7,<0.3.0
yfinance>=0.2.36,<0.3.0
matplotlib>=3.7.0,<4.0.0
plotly>=5.18.0,<6.0.0 